separate ops
