+++ src/symbols/enum_based.rs
@@ -96,128 +96,3 @@ where
         Box::new((*self).diff(dm))
     }
 }
-
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub enum BinaryOpSet {
-    Add,
-    Sub,
-    Mul,
-    Div,
-    Pow,
-}
-
-impl BinaryOp for BinaryOpSet {}
-
-impl<Sym1, Sym2, Out, In: ?Sized> DynamicSymbol<Out, In>
-    for BinarySym<BinaryOpSet, Sym1, Sym2, Out, In>
-where
-    Sym1: Symbol<Out, In>,
-    Sym2: Symbol<Out, In>,
-    Out: BasicNumOps + Any,
-    In: Any,
-{
-    fn calc_ref(&self, i: &In) -> Out {
-        match &self.op {
-            BinaryOpSet::Add => {
-                BinarySym::new_with_op(AddOp, self.sym1.clone(), self.sym2.clone()).calc_ref(i)
-            }
-            BinaryOpSet::Sub => {
-                BinarySym::new_with_op(SubOp, self.sym1.clone(), self.sym2.clone()).calc_ref(i)
-            }
-            BinaryOpSet::Mul => {
-                BinarySym::new_with_op(MulOp, self.sym1.clone(), self.sym2.clone()).calc_ref(i)
-            }
-            BinaryOpSet::Div => {
-                BinarySym::new_with_op(DivOp, self.sym1.clone(), self.sym2.clone()).calc_ref(i)
-            }
-            _ => todo!(),
-        }
-    }
-    fn diff_dyn(&self, dm: usize) -> DynExpr<Out, In> {
-        todo!()
-    }
-    fn as_any(&self) -> &(dyn Any) {
-        todo!()
-    }
-}
-
-impl<Sym1, Sym2, Out, In: ?Sized> Symbol<Out, In> for BinarySym<BinaryOpSet, Sym1, Sym2, Out, In>
-where
-    Sym1: Symbol<Out, In>,
-    Sym2: Symbol<Out, In>,
-    Out: BasicNumOps + Any,
-    In: Any,
-    EExpr<Out, In>: Symbol<Out, In>,
-{
-    type Derivative = EExpr<Out, In>;
-    fn diff(self, _: usize) -> <Self as Symbol<Out, In>>::Derivative {
-        todo!()
-    }
-}
-
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub enum UnaryOpSet {
-    Neg,
-    Square,
-    FloatOp(FloatOpSet),
-}
-
-#[derive(Clone, PartialEq, Eq, Hash, Debug)]
-pub enum FloatOpSet {}
-
-impl UnaryOp for UnaryOpSet {
-    fn op_name<'a>(&self) -> &'a str {
-        match self {
-            UnaryOpSet::Neg => NegOp.op_name(),
-            UnaryOpSet::Square => SquareOp.op_name(),
-            _ => unimplemented!(),
-        }
-    }
-
-    ///Formats the expression to display.
-    fn format_expression<Out, In: ?Sized>(
-        &self,
-        f: &mut fmt::Formatter<'_>,
-        inner: &impl DynamicSymbol<Out, In>,
-    ) -> Result<(), fmt::Error> {
-        match self {
-            UnaryOpSet::Neg => NegOp.format_expression(f, inner),
-            UnaryOpSet::Square => SquareOp.format_expression(f, inner),
-            _ => unimplemented!(),
-        }
-    }
-}
-
-impl<Sym, Out, In: ?Sized> DynamicSymbol<Out, In> for UnarySym<UnaryOpSet, Sym, Out, In>
-where
-    Sym: Symbol<Out, In>,
-    Out: BasicNumOps + Any,
-    In: Any,
-{
-    fn calc_ref(&self, i: &In) -> Out {
-        match &self.op {
-            UnaryOpSet::Neg => UnarySym::new_with_op(NegOp, self.sym.clone()).calc_ref(i),
-            UnaryOpSet::Square => UnarySym::new_with_op(SquareOp, self.sym.clone()).calc_ref(i),
-            UnaryOpSet::FloatOp(fop) => todo!(),
-        }
-    }
-    fn diff_dyn(&self, dm: usize) -> symbols::dyn_expr::DynExpr<Out, In> {
-        todo!()
-    }
-    fn as_any(&self) -> &(dyn std::any::Any) {
-        todo!()
-    }
-}
-
-impl<Sym, Out, In: ?Sized> Symbol<Out, In> for UnarySym<UnaryOpSet, Sym, Out, In>
-where
-    Sym: Symbol<Out, In>,
-    Out: BasicNumOps + Any,
-    In: Any,
-    EExpr<Out, In>: Symbol<Out, In>,
-{
-    type Derivative = EExpr<Out, In>;
-    fn diff(self, _: usize) -> <Self as Symbol<Out, In>>::Derivative {
-        todo!()
-    }
-}
